refactor: reorganize filter and sort constructor
